SCB  session、window、mission ID归一处理

+++ window_scene/common/include/window_session_property.h
@@ -69,8 +69,8 @@ public:
     void SetDisplayId(uint64_t displayId);
     void SetWindowType(WindowType type);
     void SetParentId(uint32_t parentId);
-    void SetPersistentId(uint64_t persistentId);
-    void SetParentPersistentId(uint64_t persistentId);
+    void SetPersistentId(uint32_t persistentId);
+    void SetParentPersistentId(uint32_t persistentId);
     void SetAccessTokenId(uint32_t accessTokenId);
     void SetTokenState(bool hasToken);
     void SetMaximizeMode(MaximizeMode mode);
@@ -101,8 +101,8 @@ public:
     uint32_t GetParentId() const;
     uint32_t GetWindowFlags() const;
     uint64_t GetDisplayId() const;
-    uint64_t GetPersistentId() const;
-    uint64_t GetParentPersistentId() const;
+    uint32_t GetPersistentId() const;
+    uint32_t GetParentPersistentId() const;
     uint32_t GetAccessTokenId() const;
     bool GetTokenState() const;
     MaximizeMode GetMaximizeMode() const;
@@ -140,8 +140,8 @@ private:
     uint64_t displayId_ = 0;
     uint32_t parentId_ = INVALID_SESSION_ID; // parentId of sceneSession, which is low 32 bite of parentPersistentId_
     uint32_t flags_ = 0;
-    uint64_t persistentId_ = INVALID_SESSION_ID;
-    uint64_t parentPersistentId_ = INVALID_SESSION_ID;
+    uint32_t persistentId_ = INVALID_SESSION_ID;
+    uint32_t parentPersistentId_ = INVALID_SESSION_ID;
     uint32_t accessTokenId_ = INVALID_SESSION_ID;
     MaximizeMode maximizeMode_ = MaximizeMode::MODE_RECOVER;
     WindowMode windowMode_ = WindowMode::WINDOW_MODE_FULLSCREEN;
