+++ window_scene/session_manager/include/session_manager.h
@@ -30,10 +30,10 @@ WM_DECLARE_SINGLE_INSTANCE_BASE(SessionManager);
 public:
     void CreateAndConnectSpecificSession(const sptr<ISessionStage>& sessionStage,
         const sptr<IWindowEventChannel>& eventChannel, const std::shared_ptr<RSSurfaceNode>& surfaceNode,
-        sptr<WindowSessionProperty> property, uint64_t& persistentId, sptr<ISession>& session);
-    void DestroyAndDisconnectSpecificSession(const uint64_t& persistentId);
+        sptr<WindowSessionProperty> property, uint32_t& persistentId, sptr<ISession>& session);
+    void DestroyAndDisconnectSpecificSession(const uint32_t& persistentId);
     WMError UpdateProperty(sptr<WindowSessionProperty>& property, WSPropertyChangeAction action);
-    WMError SetSessionGravity(uint64_t persistentId, SessionGravity gravity, uint32_t percent);
+    WMError SetSessionGravity(uint32_t persistentId, SessionGravity gravity, uint32_t percent);
     WMError BindDialogTarget(uint64_t persistentId, sptr<IRemoteObject> targetToken);
 
     sptr<ISceneSessionManager> GetSceneSessionManagerProxy();
